Make multimethods work.
Did a small amount of cleanup. Tested this version with monolith,
and it seems to work like a dream.

Only problem is it broke with every other cljs environment. It looks
like I'm going to have to allow for different implementations for
loading namespaces. The browser will be inherently async, and the
post-operations (state+after) will need to occur in a callback. Node
namespace loading is synchronous, and the post-operations need to occur
immediately. Same for nashorn. Additionally, the loading functions
change from environment to environment (goog.net.jsloader in browser,
js/require in node, js/load in nashorn).

I think a decent solution would be to dispatch loading and
post-operations on repl-env type. Unforunately, these will need to be
tied together in the same function (because some are async and some are
sync, and the post-operations *must* happen after loading).

After thinking it through, it would probably be best to dispatch the
whole process on repl-env type, and use code sharing for the the
implementation bits that can be shared.